Remove feedback modal from CYS flow (#59944)
Since we decided to move on from CYS and our focus is not going to be
there anytime soon, we decided to remove the “Share feedback” button
at the end of the flow. For the past months, we haven't been reading
any feedback, and it'd be better for both us and our users not to collect
unactionable feedback: we want users to feel heard when they take
the time to submit feedback, and we only want to collect feedback about
our actual focuses.

Decision was made in tandem with @nerrad and @SamratBiswas1

Closes #52283