Email Editor: Enable Site Logo and Site Title Blocks in the Email Editor (#59624)
* Disable duotone support from the email editor core.

* Update email template to use site logo block instead of static image for improved flexibility

* Enable site logo and site title blocks for use in the email editor

* Add changelog file

* Remove default underline from site title link in email editor styles

This is mainly to maintain style consistency between the email editor and the rendered output

* Refactor email template to dynamically include site logo or title

Updated the WooEmailTemplate class to replace the static site logo with a dynamic method that checks for the presence of a custom logo. This enhances flexibility in email templates by allowing the display of either the site logo or title based on availability.

* Add Woo plugin changelog file

* Apply quick PR review feedback

* Ensure we fetch image size with the correct parameter.

This fixes the warnings: Failed to open stream: Connection refused in ...

* Ensure Site Title and Site Logo blocks renders correctly

* Code style lint fixes

* Enhance Site Logo block by adding a custom variation for the email editor. This includes registering a default centered alignment and width for improved presentation.

* Add JS package changelog file