Switch to prefixed version of the CssInliner package (#61191)
* Remove pelago/emogrifier dependency from email-editor and WooCommerce composer files, updating composer.lock accordingly.

We need to remove the pelago/emogrifier package, e.g, (composer remove pelago/emogrifier) from both the WooCommerce plugin and the Email Editor because removing it from only the plugin will still let it be included via the email editor package.
Since both the email editor package and the plugin rely on the same package, we need to update both simultaneously.

* Add Mozart config for the emogrifier package.

The `thecodingmachine/safe` config is required for `pelago/emogrifier` 8.1.0. This is required package install or update will throw an error without it.

* Add pelago/emogrifier to the Woo plugin and prefix the classes

* Use Prefixed package versions in the Woo core plugin and the email editor

* Add changelog files

* Patch phpstan errors

* Fix lint errors
The proper type will be added in another PR.

* Skip email editor tests pending when the final PR is completed.

* Skip markdownlint for generated files