Use Hex for `<` and `>` Inside `<script>` Tags (#61245)
When rendering inside of a script tag, the printing of
a <!-- comment and a </script> tag will break the
rest of the page. This is due to some very
complicated HTML parsing behavior detailed
in WordPress Trac #63851. The problem can
be avoided by escaping angle brackets and
converting them to hex. JavaScript will
parse the hex and see it as the appropriate
character while HTML will ignore it.

* Add changefile(s) from automation for the following project(s): packages/php/email-editor, woocommerce

* Style Fix

---------

Co-authored-by: github-actions <github-actions@github.com>