Email Editor - basic compatibility fixes with Gutenberg 22 (#61925)
* Fix fetching global styles post

The post can be fetched both ways,
but we need to use the same method as core does
so that both core can properly track changes.

* Fix error coming from woocommerce/add-to-cart-with-options-variation-selector-attribute

This block is missing attributes and it causes crashes.
This may need to be addressed also upstream.

* Hide style core style panel in email editor

Currently there is no API for hiding it so we need to use CSS.

* Add overrides for styleId and theme

This is a workaround to enforce email theme and styles to be loaded by core.

* Fix issues causing internal re-initialization of editor settings

* Add dynamic styles for email layout to theme styles

We can't easily add additional dynamically generated CSS as styles can no longer
be passed via Editor component property.
In this workaround we pass these styles with base theme styles.

* Skip generating styles when useGlobalStylesOutputWithConfig is missing

This is to prevent crashes, but useEmailCss will not generate styles
with Gutenberg 22.0+

* Skip store overrides if not needed

* Use dummy hook when useGlobalStylesOutputWithConfig

This is to avoid breaking conditional hooks rule.

* Fix issues reported by linter

* Add changelogs

* Add test mocks for private APIs

* Address code rabbit's remarks