Email editor Gutenberg 22/WP 6.9 compatibility fixes (#61964)
* Add custom hook for generating styles and re-enable use-email-css hook

WOOPRD-1278

* Update admin client build to bundle the @wordpress/global-styles-engine

The package is exposed via wp global variable so we need to bundle it
WOOPRD-1278

* Cleanup private APIs exports

* Update types for email theme

* Add overrides deactivation

This is needed when the editor is used in an app that allows
editing emails and also regular posts
WOOPRD-1278

* Fix appearance of styles panel

WOOPRD-1278

* Fix back button styles

The header height changed from 60 to 64px.
This fix optimizes for ideal appearance in WP6.9+
For WP 6.8 there is still small misalignment in hover state on the button.
WOOPRD-1278

* Add changelog

* Fix linter warnings

* Remove outdated comment

* Remove optional getBlockStyles from global styles hook

* Update email styles directly via editor settings property

Gutenberg 22+ supports passing styles via editor settings.
This approach is similar to what we did previously and will
allow us to get rid of the core editor store overrides, which might
be problematic because there is no 100% reliable mechanism for disabling
them.

* Remove Editor store overrides

* Update changelog

* Add a note about bundling the global styles package

* Address nitpicks, comments and fix types for global styles post