Use store title as default email "From" name instead of hardcoded WordPress site name (#62369)
* Add default fallback to blogname for get_from_name methods

Update get_from_name methods in WC_Emails and WC_Email classes to ensure consistent behavior when the email name option is not set.

Adjust TemplateApiController to reflect the new default behavior for sender settings.

* Add 'skip_initial_save' option to settings and update installation logic

When set to true, this allows the plugin to skip saving the initial option value in the db for the settings field.

In this case, we don't want to save the initial value for the woocommerce_email_from_name field. This will allow us to continue using the blog title for the email “from” until the user manually updates it or saves the email settings field, which will repopulate the option.

* Add changelog and fix lint warnings

* Lint warning fixes

* Remove strict type declarations from the email settings class for backwards compatibility purposes.

* Remove deprecated PHPDoc parse error entries from WooCommerce PHPStan baseline file to streamline linting process.